apps/api/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
sqlmodel>=0.0.21,<0.0.23
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
supabase>=2.0,<3.0
pydantic>=2.7,<3.0
